Move GetRevitParameterValues to new file
